Refactor: 프로필 페이지 코드 리팩토링 및 버그 수정 (#117)
* Feat: 비밀번호 변경 성공 후 로그아웃 추가 (#108)

* Feat: 회원탈퇴 성공 후 로그아웃, 에러 코드 분기 추가 (#108)

* Fix: 이미지 기본값 null일 때 조건 추가 (#108)

* Fix: 로그아웃 커스텀훅 공통 사용법대로 수정 (#108)

* Feat: useParams 커스텀 훅에 추가 (#108)

* Refactor: 이미지 미리보기 함수 수정 및 이미지 관련 주석 추가 (#108)

* Feat: 회원 탈퇴 전 확인창 추가 (#108)

* Refactor: VAC 패턴으로 정리 (#108)

+++ client/src/pages/Profile/DeleteUser.tsx
@@ -1,12 +1,14 @@
 import { useState } from 'react'
 import { useRecoilValue } from 'recoil'
 import { userInfo } from '../../store/store'
+import useLogout from '../../hooks/useLogout'
 import useRouter from '../../hooks/useRouter'
 import { deleteUser } from '../../apis/user'
 import styles from './DeleteUser.module.scss'
 
 function DeleteUser() {
   const { routeTo } = useRouter()
+  const logout = useLogout()
   const { userId } = useRecoilValue(userInfo)
   const [currentPassword, setCurrentPassword] = useState('')
 
@@ -16,15 +18,21 @@ function DeleteUser() {
 
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault()
+
+    const shouldDelete = window.confirm('정말로 탈퇴하시겠습니까?')
+    if (!shouldDelete) {
+      return
+    }
+
     deleteUser({ userId: userId, password: currentPassword }).then(response => {
       if (response?.status === 200) {
         alert('회원탈퇴가 완료되었습니다.')
-        // 체크박스 추가
-        // todo: 로그아웃 처리
+        logout()
         routeTo('/')
-      }
-      if (response?.status !== 200) {
+      } else if (response?.status === 400) {
         alert('비밀번호가 일치하지 않습니다.')
+      } else {
+        alert('회원 탈퇴에 실패했습니다. 자세한 내용은 관리자에게 문의해 주시기 바랍니다.')
       }
     })
   }
